sqlite conversion
- make sqlite the default for dev & test environments
- remove default db configs and rely on environment variables
- move db dialect specific configs to separate const

+++ tests/globalTeardown.js
@@ -10,7 +10,7 @@ module.exports = async () => {
     const database = new Database();
 
     const isDbAvailable = await database.isAvailable();
-    if (isDbAvailable && database.isSqlite()) {
+    if (isDbAvailable && database.isSqlite() && database.config.storage) {
         debug(`removing sqlite database ${database.config.storage}`);
         fs.unlinkSync(database.config.storage);
     }
